renamed trash restore module for readability
