+++ src/actions.rs
@@ -1,3 +1,3 @@
 pub mod trash_list;
 pub mod trash_put;
-pub mod trash_rs;
+pub mod trash_restore;
